move clients to `repository` package
