+++ images/agent/internal/controller/llv/reconciler.go
@@ -38,6 +38,7 @@ import (
 	"github.com/deckhouse/sds-node-configurator/images/agent/internal/controller"
 	"github.com/deckhouse/sds-node-configurator/images/agent/internal/logger"
 	"github.com/deckhouse/sds-node-configurator/images/agent/internal/monitoring"
+	"github.com/deckhouse/sds-node-configurator/images/agent/internal/repository"
 	"github.com/deckhouse/sds-node-configurator/images/agent/internal/utils"
 )
 
@@ -58,8 +59,8 @@ type cleanups struct {
 type Reconciler struct {
 	cl              client.Client
 	log             logger.Logger
-	lvgCl           *utils.LVGClient
-	llvCl           *utils.LLVClient
+	lvgCl           *repository.LVGClient
+	llvCl           *repository.LLVClient
 	metrics         monitoring.Metrics
 	sdsCache        *cache.Cache
 	cfg             ReconcilerConfig
@@ -88,14 +89,14 @@ func NewReconciler(
 	return &Reconciler{
 		cl:  cl,
 		log: log,
-		lvgCl: utils.NewLVGClient(
+		lvgCl: repository.NewLVGClient(
 			cl,
 			log,
 			metrics,
 			cfg.NodeName,
 			ReconcilerName,
 		),
-		llvCl: utils.NewLLVClient(
+		llvCl: repository.NewLLVClient(
 			cl, log,
 		),
 		metrics:  metrics,
