+++ images/agent/internal/controller/llvs/reconciler_ee.go
@@ -30,6 +30,7 @@ import (
 	"github.com/deckhouse/sds-node-configurator/images/agent/internal/controller"
 	"github.com/deckhouse/sds-node-configurator/images/agent/internal/logger"
 	"github.com/deckhouse/sds-node-configurator/images/agent/internal/monitoring"
+	"github.com/deckhouse/sds-node-configurator/images/agent/internal/repository"
 	"github.com/deckhouse/sds-node-configurator/images/agent/internal/utils"
 )
 
@@ -38,7 +39,7 @@ const ReconcilerName = "lvm-logical-volume-snapshot-watcher-controller"
 type Reconciler struct {
 	cl       client.Client
 	log      logger.Logger
-	lvgCl    *utils.LVGClient
+	lvgCl    *repository.LVGClient
 	metrics  monitoring.Metrics
 	sdsCache *cache.Cache
 	cfg      ReconcilerConfig
@@ -63,7 +64,7 @@ func NewReconciler(
 	return &Reconciler{
 		cl:  cl,
 		log: log,
-		lvgCl: utils.NewLVGClient(
+		lvgCl: repository.NewLVGClient(
 			cl,
 			log,
 			metrics,
